text-controller-mvp/src/app/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pyserial==3.5
requests==2.32.5
urllib3==2.6.3
